Add extra parameters to the DockerClient constructor's
The defaults remain 10 seconds for the read time out and to enable the
LoggingFilter.

Allows the caller to change the read time out and to disable the LoggingFilter

For large images the logging filter spam's the console with the binary data that
is being transferred.

This new setting allows it to be turned off.