Adds test for CopyFileFromContainer
This changes the CopyFileFromContainer's impl method to not call
toString on the copyConfig that it uses in the POST body, instead
relying on the default JSON serializer and marked fields. This will
allow us to use a reflection-based toString.

This commit also creates a test around the use of the
CopyFileFromContainer command.