Backport the new structure to Jersey 1.18
This ressurects the previous Jersey 1.18 implementation in the new
structure.  The Jersey 2.x implementation is still here, but has been
moved to the jaxrs2 package which currently is not built.  Eventually
these could be split into separate modules which could be combined as
desired, but there is still a bit more work to do for that to work.