Use canonical form of Docker folder when building TAR files
In BuildImageCmdImpl#buildDockerFolderTar a canonical form of the source
files referenced in the Dockerfile is used. But the Docker folder is
passed in the given form to the CompressArchiveUtil.

CompressArchiveUtil#relativize creates absolute TAR archive entries if the
canonical form of the source files differs from the given form of the
Docker folder. As a result, the Docker deamon can not find the files
during the build. This can happen in case-insensitive file systems such as
Windows, for example.

As a solution, the canonical form of the Docker folder must be used.