Do not serialize unconfigured values in StartContainerCmd
Any payload sent with the StartContainerCmd will cause the target
container to lose its customization.
So in order to just start a preconfigured container you must be able to
send an empty "{}" message to the /containers/{id}/start endpoint.

This is accomplished by all configuration defaulting to null in
combination with @JsonInclude(NON_EMPTY).