Do not wait for another event before exiting.
It the callback decideds it is no longer interested in receiving events
whilst it handles either an event or an exception then it should not have to
wait until the next event from Docker is received to exit the loop.

This adds a check before the event is received and leaves the check after
the event has received as the event may take a long time to come and the
callback may have decided during that time that it wants to shutdown.