Stop leaking tar files in temporary folder
The temporary tar files generated from the docker folder are preserved after
the contents is send to the docker deamon. In order to prevent filling up the
temporary directory those files must be deleted.

As the API is based on InputStreams the easiest solution is to wrap the input
stream an delete the file on close.