Add API versioning support
An issue with docker-java currently is that API < 1.19 has a different
authentication scheme than newer API versions. This leads to major problems
in environments with different docker installations
(see alexec/docker-java-orchestration#49)

docker-java Issue #288 fixed the issue for API >= 1.19, but let older servers
out of the consideration.

This patch distributes the client config through the executors and thus allows
the execution to vary depending on those options.