Allow legacy registry message to indicate pull success
If build node is pulling from older registry and build node reports
legacy registry message, docker java will treat this as a failure.
This update allows this message to be treated as a success condition.

Add tests for PullItemResponse

Add new tests for testing the error/success logic of
PullItemResponse. Update test framework with minor refactoring.