Always use HTTP to connect to proxy even when proxying HTTPS
This corresponds with how the java proxy properties work as well. Even
when proxying HTTPS the connection to the actually proxy is done over
HTTP and it makes a CONNECT call to stream the HTTPS over.

Most proxies are still used via HTTP and for better or worse this is
the assumption built into the java proxy configuration which
docker-java is emulating. The apache client is more flexible and
perhaps the docker-java configuration should be too but I'd like to
argue that until then that the docker-java match the behavior of java.

Also this allows me to use docker-java through the corporate proxy
when proxying HTTPS. Otherwise HTTP works but HTTPS fails since the
proxy is not configured with HTTPS.