Fix CompressArchiveUtil unit tests
The tar method's contract is very clear, it creates a tar archive from
files in a folder. The unit tests should validate that the tar archive
was created and that the content is what it is supposed to be. Same
apply to the archiveTARFiles method.

Before, in order to validate that the tar archive was created, the unit
tests were un-archiving the tar. This is wrong as you end up implicitly
testing the extraction code which could be faulty.

In addition, the 2 tests for symlinks were not even checking if the
extracted file was a symlink, they were only checking if the file or
folder were readable.

Rewrite the tests to assert directly the content of the archive. Also
fix the 2 symlinks tests to make sure the tar preserve them. Fixing the
unit tests exposed that CompressArchiveUtil  methods do not preserve the
symlinks which is a bug. Add @Ignore annotation for those tests until
the bug is fixed.

This change is a preparation change to fix the archive creation to
preserve symlinks so add tests to increase archive creation code
coverage close to 100%. This will allow to fix the code while making
sure no regression are introduced.

Related to issue https://github.com/docker-java/docker-java/issues/532