Expose httpclient5 connection keepalive
Why:
We have been experiencing problems with podman and testcontainers using a
globally cached ZerodepDockerHttpClient. The problem is, with it's default configuration,
podman closes idle connections after 10 seconds. The apache HTTP client's default
connection keepalive is 3 minutes. So we are seeing exceptions thrown when the client
attempts to use these stale connections that it has in the connection pool.

A configuration that could help are reducing the connectionKeepAlive, so we could
configure the pool to close connections in line with this podman timeout.

Note: stale connection checking does not work and blocks in the isStale see #1726

https://github.com/testcontainers/testcontainers-java/issues/7310